mixins improvements, close #103
