+++ lib/browser_core.js
@@ -6,4 +6,9 @@ module.exports = class BrowserCore {
         this.page = page;
         this._settings = settings;
     }
+
+    evaluate(cb, ...args) {
+        this._failIfNotLoaded();
+        return this.page.evaluate(cb, ...args);
+    }
 };
