+++ lib/factory/browser_factory.js
@@ -7,8 +7,7 @@ const FactoryCore = require('./factory_core');
 
 const mixins = [
     require("../mixins/browser_actions"),
-    require("../mixins/browser_base_mixin"),
-    require("../mixins/browser_mixin"),
+    require("../mixins/browser_info"),
     require("../mixins/browser_navigation"),
     require("../mixins/browser_queries"),
     require("../mixins/browser_setup"),
