+++ lib/mixins/browser_queries.js
@@ -3,11 +3,6 @@
 module.exports = function BrowserQueriesMixin(s) {
     return class BrowserQueries extends s {
 
-        evaluate(cb, ...args) {
-            this._failIfNotLoaded();
-            return this.page.evaluate(cb, ...args);
-        }
-
         query(selector, optionalSelector) {
             this._failIfNotLoaded();
             if(optionalSelector) {
